Core/Movement: Changed movement flags2 to uint32

+++ src/server/game/Entities/Object/Object.h
@@ -362,7 +362,7 @@ struct MovementInfo
     // common
     ObjectGuid guid;
     uint32 flags;
-    uint16 flags2;
+    uint32 flags2;
     Position pos;
     uint32 time;
 
@@ -422,11 +422,11 @@ struct MovementInfo
     void RemoveMovementFlag(uint32 flag) { flags &= ~flag; }
     bool HasMovementFlag(uint32 flag) const { return (flags & flag) != 0; }
 
-    uint16 GetExtraMovementFlags() const { return flags2; }
-    void SetExtraMovementFlags(uint16 flag) { flags2 = flag; }
-    void AddExtraMovementFlag(uint16 flag) { flags2 |= flag; }
-    void RemoveExtraMovementFlag(uint16 flag) { flags2 &= ~flag; }
-    bool HasExtraMovementFlag(uint16 flag) const { return (flags2 & flag) != 0; }
+    uint32 GetExtraMovementFlags() const { return flags2; }
+    void SetExtraMovementFlags(uint32 flag) { flags2 = flag; }
+    void AddExtraMovementFlag(uint32 flag) { flags2 |= flag; }
+    void RemoveExtraMovementFlag(uint32 flag) { flags2 &= ~flag; }
+    bool HasExtraMovementFlag(uint32 flag) const { return (flags2 & flag) != 0; }
 
     uint32 GetFallTime() const { return jump.fallTime; }
     void SetFallTime(uint32 fallTime) { jump.fallTime = fallTime; }
